scripts: respect an initial LD_LIBRARY_PATH with env_use_destdir
In the highly unlikely that you would need to invoke a tool with
an out of tree PCRE2 library, it comes handy to set LD_LIBRARY_PATH
to point to such library and not get your setting blown away.

As usual, if no LD_LIBRARY_PATH exists previous to calling the
script, nothing will change.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>